+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Task display: Use cooperative cancellation for cancel buttons in task display.
 - Task display: Print task progress every 5 seconds for 'plain' display mode.
 - Task display: Handle click on running samples tab when there is no transcript.
+- Docker: Print sample id and epoch for each container when using `--no-sandbox-cleanup`
 - Mistral: Create and destroy client within generate.
 - Inspect View: Fix display of score dictionaries containing boolean values
 - Bugfix: Catch standard `TimeoutError` for subprocess timeouts (ensure kill/cleanup of timed out process).